+++ CDF3/src/libf/get_var_intf.c
@@ -21,7 +21,7 @@
 
 /* Prototypes for the Fortran interfaces */
 #include "mpifnetcdf.h"
-FORTRAN_API int FORT_CALL nfmpi_get_var_int_ ( int *v1, int64_t *v2, MPI_Fint *v3 ){
+FORTRAN_API int FORT_CALL nfmpi_get_var_int_ ( int *v1, int *v2, MPI_Fint *v3 ){
     int ierr;
     int l2 = *v2 - 1;
     ierr = ncmpi_get_var_int( *v1, l2, v3 );
