+++ CDF3/src/libf/put_var_int2f.c
@@ -21,7 +21,7 @@
 
 /* Prototypes for the Fortran interfaces */
 #include "mpifnetcdf.h"
-FORTRAN_API int FORT_CALL nfmpi_put_var_int2_ ( int *v1, int64_t *v2, short*v3 ){
+FORTRAN_API int FORT_CALL nfmpi_put_var_int2_ ( int *v1, int *v2, short*v3 ){
     int ierr;
     int l2 = *v2 - 1;
     ierr = ncmpi_put_var_short( *v1, l2, v3 );
